back-end/backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
python-dotenv
google-genai
pymongo
